refactor: co-locate skeleton components with their real counterparts
Move each skeleton out of the monolithic common/Skeletons.tsx into the
same directory as the component it mirrors, so layout changes are
visible in the same folder.

Extract shared layout constants (FEED_CARD_SX, FEED_IMAGE_MAX_HEIGHT,
PROFILE_HEADER_SX, PROFILE_STAT_BOX_SX, PROFILE_AVATAR_SIZE) into
feedLayout.ts and profileLayout.ts so skeleton and real component
reference the same values instead of duplicating magic numbers.
