+++ frontend/src/components/feed/FeedView.tsx
@@ -17,7 +17,8 @@ import { loadFeed, loadInitialFeed, switchTab } from "../../store/feedSlice";
 import { openEditModal, openDeleteConfirm } from "../../store/uiSlice";
 import type { FeedTab, Occurrence } from "../../services/types";
 import { FeedItem } from "./FeedItem";
-import { FeedSkeletonList, ProfileObservationCardSkeleton } from "../common/Skeletons";
+import { FeedSkeletonList } from "./FeedItemSkeleton";
+import { ProfileObservationCardSkeleton } from "../profile/ProfileObservationCardSkeleton";
 import { ExploreFilterPanel } from "./ExploreFilterPanel";
 import { getImageUrl } from "../../services/api";
 import { formatTimeAgo, getObservationUrl } from "../../lib/utils";
